OLED menu 95% complete

+++ gruppe6/gruppe6/rs232_driver.h
@@ -3,7 +3,7 @@
 #include <avr/io.h>
 #include <stdio.h>
 
-void UART_Init( unsigned int ubrr);
+void UART_Init( uint16_t ubrr);
 void UART_Transmit( unsigned char data );
 unsigned char UART_Receive( void);
